+++ packages/shared/src/modals/TransactionSnackbar/index.tsx
@@ -9,7 +9,10 @@ export interface TransactionSnackbarProps {
     pluginID: NetworkPluginID
 }
 
-export const TransactionSnackbarModal = forwardRef<SingletonModalRefCreator, TransactionSnackbarProps>((props, ref) => {
+export const TransactionSnackbarModal = forwardRef<
+    SingletonModalRefCreator<TransactionSnackbarOpenProps>,
+    TransactionSnackbarProps
+>((props, ref) => {
     useSingletonModal(ref)
 
     return <TransactionSnackbar pluginID={props.pluginID} />
